Reorganize domain model and execution messages
