+++ crates/common/src/messages/execution/cancel.rs
@@ -13,17 +13,14 @@
 //  limitations under the License.
 // -------------------------------------------------------------------------------------------------
 
-use std::{cell::RefCell, fmt::Display, rc::Rc};
+use std::fmt::Display;
 
 use derive_builder::Builder;
 use nautilus_core::{UUID4, UnixNanos};
-use nautilus_model::{
-    identifiers::{ClientId, ClientOrderId, InstrumentId, StrategyId, TraderId, VenueOrderId},
-    orders::OrderAny,
+use nautilus_model::identifiers::{
+    ClientId, ClientOrderId, InstrumentId, StrategyId, TraderId, VenueOrderId,
 };
 use serde::{Deserialize, Serialize};
-
-use crate::order_emulator::emulator::OrderEmulator;
 
 #[derive(Clone, PartialEq, Eq, Debug, Default, Serialize, Deserialize, Builder)]
 #[builder(default)]
@@ -76,25 +73,6 @@ impl Display for CancelOrder {
             "CancelOrder(instrument_id={}, client_order_id={}, venue_order_id={})",
             self.instrument_id, self.client_order_id, self.venue_order_id,
         )
-    }
-}
-
-pub trait CancelOrderHandler {
-    fn handle_cancel_order(&self, order: &OrderAny);
-}
-
-#[derive(Debug)]
-pub enum CancelOrderHandlerAny {
-    OrderEmulator(Rc<RefCell<OrderEmulator>>),
-}
-
-impl CancelOrderHandler for CancelOrderHandlerAny {
-    fn handle_cancel_order(&self, order: &OrderAny) {
-        match self {
-            Self::OrderEmulator(order_emulator) => {
-                order_emulator.borrow_mut().cancel_order(order);
-            }
-        }
     }
 }
