+++ crates/execution/src/client/mod.rs
@@ -13,24 +13,18 @@
 //  limitations under the License.
 // -------------------------------------------------------------------------------------------------
 
+use nautilus_common::messages::execution::{
+    BatchCancelOrders, CancelAllOrders, CancelOrder, ModifyOrder, QueryOrder, SubmitOrder,
+    SubmitOrderList,
+    reports::{GenerateFillReports, GenerateOrderStatusReport, GeneratePositionReports},
+};
 use nautilus_core::UnixNanos;
 use nautilus_model::{
     accounts::AccountAny,
     enums::OmsType,
     identifiers::{AccountId, ClientId, Venue},
+    reports::{ExecutionMassStatus, FillReport, OrderStatusReport, PositionStatusReport},
     types::{AccountBalance, MarginBalance},
-};
-
-use crate::{
-    messages::{
-        BatchCancelOrders, CancelAllOrders, CancelOrder, ModifyOrder, QueryOrder, SubmitOrder,
-        SubmitOrderList,
-        reports::{GenerateFillReports, GenerateOrderStatusReport, GeneratePositionReports},
-    },
-    reports::{
-        fill::FillReport, mass_status::ExecutionMassStatus, order::OrderStatusReport,
-        position::PositionStatusReport,
-    },
 };
 
 pub mod base;
@@ -42,6 +36,7 @@ pub trait ExecutionClient {
     fn venue(&self) -> Venue;
     fn oms_type(&self) -> OmsType;
     fn get_account(&self) -> Option<AccountAny>;
+
     /// Generates and publishes the account state event.
     ///
     /// # Errors
@@ -54,54 +49,63 @@ pub trait ExecutionClient {
         reported: bool,
         ts_event: UnixNanos,
     ) -> anyhow::Result<()>;
+
     /// Starts the execution client.
     ///
     /// # Errors
     ///
     /// Returns an error if the client fails to start.
     fn start(&mut self) -> anyhow::Result<()>;
+
     /// Stops the execution client.
     ///
     /// # Errors
     ///
     /// Returns an error if the client fails to stop.
     fn stop(&mut self) -> anyhow::Result<()>;
+
     /// Submits a single order command to the execution venue.
     ///
     /// # Errors
     ///
     /// Returns an error if submission fails.
     fn submit_order(&self, cmd: &SubmitOrder) -> anyhow::Result<()>;
+
     /// Submits a list of orders to the execution venue.
     ///
     /// # Errors
     ///
     /// Returns an error if submission fails.
     fn submit_order_list(&self, cmd: &SubmitOrderList) -> anyhow::Result<()>;
+
     /// Modifies an existing order.
     ///
     /// # Errors
     ///
     /// Returns an error if modification fails.
     fn modify_order(&self, cmd: &ModifyOrder) -> anyhow::Result<()>;
+
     /// Cancels a specific order.
     ///
     /// # Errors
     ///
     /// Returns an error if cancellation fails.
     fn cancel_order(&self, cmd: &CancelOrder) -> anyhow::Result<()>;
+
     /// Cancels all orders.
     ///
     /// # Errors
     ///
     /// Returns an error if cancellation fails.
     fn cancel_all_orders(&self, cmd: &CancelAllOrders) -> anyhow::Result<()>;
+
     /// Cancels a batch of orders.
     ///
     /// # Errors
     ///
     /// Returns an error if batch cancellation fails.
     fn batch_cancel_orders(&self, cmd: &BatchCancelOrders) -> anyhow::Result<()>;
+
     /// Queries the status of an order.
     ///
     /// # Errors
@@ -117,12 +121,14 @@ pub trait LiveExecutionClient: ExecutionClient {
     ///
     /// Returns an error if connection fails.
     fn connect(&self) -> anyhow::Result<()>;
+
     /// Disconnects the live execution client.
     ///
     /// # Errors
     ///
     /// Returns an error if disconnection fails.
     fn disconnect(&self) -> anyhow::Result<()>;
+
     /// Generates a single order status report.
     ///
     /// # Errors
@@ -132,6 +138,7 @@ pub trait LiveExecutionClient: ExecutionClient {
         &self,
         cmd: &GenerateOrderStatusReport,
     ) -> anyhow::Result<Option<OrderStatusReport>>;
+
     /// Generates multiple order status reports.
     ///
     /// # Errors
@@ -141,6 +148,7 @@ pub trait LiveExecutionClient: ExecutionClient {
         &self,
         cmd: &GenerateOrderStatusReport,
     ) -> anyhow::Result<Vec<OrderStatusReport>>;
+
     /// Generates fill reports based on execution results.
     ///
     /// # Errors
@@ -148,6 +156,7 @@ pub trait LiveExecutionClient: ExecutionClient {
     /// Returns an error if fill report generation fails.
     fn generate_fill_reports(&self, report: GenerateFillReports)
     -> anyhow::Result<Vec<FillReport>>;
+
     /// Generates position status reports.
     ///
     /// # Errors
@@ -157,6 +166,7 @@ pub trait LiveExecutionClient: ExecutionClient {
         &self,
         cmd: &GeneratePositionReports,
     ) -> anyhow::Result<Vec<PositionStatusReport>>;
+
     /// Generates mass status for executions.
     ///
     /// # Errors
