+++ crates/execution/src/lib.rs
@@ -40,12 +40,7 @@ pub mod client;
 pub mod engine;
 pub mod matching_core;
 pub mod matching_engine;
-pub mod messages;
 pub mod models;
 pub mod order_emulator;
 pub mod order_manager;
-pub mod reports;
 pub mod trailing;
-
-#[cfg(feature = "python")]
-pub mod python;
