+++ crates/execution/src/order_emulator/adapter.rs
@@ -26,14 +26,9 @@ use nautilus_common::{
 use nautilus_core::UUID4;
 use ustr::Ustr;
 
-use crate::{
-    messages::{
-        cancel::CancelOrderHandlerAny, modify::ModifyOrderHandlerAny, submit::SubmitOrderHandlerAny,
-    },
-    order_emulator::{
-        emulator::OrderEmulator,
-        handlers::{OrderEmulatorExecuteHandler, OrderEmulatorOnEventHandler},
-    },
+use crate::order_emulator::{
+    emulator::OrderEmulator,
+    handlers::{OrderEmulatorExecuteHandler, OrderEmulatorOnEventHandler},
 };
 
 #[derive(Debug)]
@@ -47,27 +42,28 @@ impl OrderEmulatorAdapter {
 
         Self::initialize_execute_handler(emulator.clone());
         Self::initialize_on_event_handler(emulator.clone());
-        Self::initialize_submit_order_handler(emulator.clone());
-        Self::initialize_cancel_order_handler(emulator.clone());
-        Self::initialize_modify_order_handler(emulator.clone());
+        // Self::initialize_submit_order_handler(emulator.clone());
+        // Self::initialize_cancel_order_handler(emulator.clone());
+        // Self::initialize_modify_order_handler(emulator.clone());
 
         Self { emulator }
     }
 
-    fn initialize_submit_order_handler(emulator: Rc<RefCell<OrderEmulator>>) {
-        let handler = SubmitOrderHandlerAny::OrderEmulator(emulator.clone());
-        emulator.borrow_mut().set_submit_order_handler(handler);
-    }
-
-    fn initialize_cancel_order_handler(emulator: Rc<RefCell<OrderEmulator>>) {
-        let handler = CancelOrderHandlerAny::OrderEmulator(emulator.clone());
-        emulator.borrow_mut().set_cancel_order_handler(handler);
-    }
-
-    fn initialize_modify_order_handler(emulator: Rc<RefCell<OrderEmulator>>) {
-        let handler = ModifyOrderHandlerAny::OrderEmulator(emulator.clone());
-        emulator.borrow_mut().set_modify_order_handler(handler);
-    }
+    // TODO: WIP: Revisit with actor framework
+    // fn initialize_submit_order_handler(emulator: Rc<RefCell<OrderEmulator>>) {
+    //     let handler = SubmitOrderHandlerAny::OrderEmulator(emulator.clone());
+    //     emulator.borrow_mut().set_submit_order_handler(handler);
+    // }
+    //
+    // fn initialize_cancel_order_handler(emulator: Rc<RefCell<OrderEmulator>>) {
+    //     let handler = CancelOrderHandlerAny::OrderEmulator(emulator.clone());
+    //     emulator.borrow_mut().set_cancel_order_handler(handler);
+    // }
+    //
+    // fn initialize_modify_order_handler(emulator: Rc<RefCell<OrderEmulator>>) {
+    //     let handler = ModifyOrderHandlerAny::OrderEmulator(emulator.clone());
+    //     emulator.borrow_mut().set_modify_order_handler(handler);
+    // }
 
     fn initialize_execute_handler(emulator: Rc<RefCell<OrderEmulator>>) {
         let handler = ShareableMessageHandler(Rc::new(OrderEmulatorExecuteHandler {
