+++ crates/execution/src/order_emulator/emulator.rs
@@ -24,10 +24,10 @@ use nautilus_common::{
     cache::Cache,
     clock::Clock,
     logging::{CMD, EVT, RECV},
-    msgbus::{
-        handler::ShareableMessageHandler,
-        {self},
+    messages::execution::{
+        CancelAllOrders, CancelOrder, ModifyOrder, SubmitOrder, SubmitOrderList, TradingCommand,
     },
+    msgbus::{self, handler::ShareableMessageHandler},
 };
 use nautilus_core::uuid::UUID4;
 use nautilus_model::{
@@ -41,13 +41,7 @@ use nautilus_model::{
 };
 
 use crate::{
-    matching_core::OrderMatchingCore,
-    messages::{
-        CancelAllOrders, CancelOrder, ModifyOrder, SubmitOrder, SubmitOrderList, TradingCommand,
-        cancel::CancelOrderHandlerAny, modify::ModifyOrderHandlerAny,
-        submit::SubmitOrderHandlerAny,
-    },
-    order_manager::manager::OrderManager,
+    matching_core::OrderMatchingCore, order_manager::manager::OrderManager,
     trailing::trailing_stop_calculate,
 };
 
@@ -78,8 +72,7 @@ impl OrderEmulator {
         // self.register_base(portfolio, msgbus, cache, clock);
 
         let active_local = true;
-        let manager =
-            OrderManager::new(clock.clone(), cache.clone(), active_local, None, None, None);
+        let manager = OrderManager::new(clock.clone(), cache.clone(), active_local);
 
         Self {
             clock,
@@ -98,17 +91,18 @@ impl OrderEmulator {
         self.on_event_handler = Some(handler);
     }
 
-    pub fn set_submit_order_handler(&mut self, handler: SubmitOrderHandlerAny) {
-        self.manager.set_submit_order_handler(handler);
-    }
-
-    pub fn set_cancel_order_handler(&mut self, handler: CancelOrderHandlerAny) {
-        self.manager.set_cancel_order_handler(handler);
-    }
-
-    pub fn set_modify_order_handler(&mut self, handler: ModifyOrderHandlerAny) {
-        self.manager.set_modify_order_handler(handler);
-    }
+    // TODO: WIP
+    // pub fn set_submit_order_handler(&mut self, handler: SubmitOrderHandlerAny) {
+    //     self.manager.set_submit_order_handler(handler);
+    // }
+    //
+    // pub fn set_cancel_order_handler(&mut self, handler: CancelOrderHandlerAny) {
+    //     self.manager.set_cancel_order_handler(handler);
+    // }
+    //
+    // pub fn set_modify_order_handler(&mut self, handler: ModifyOrderHandlerAny) {
+    //     self.manager.set_modify_order_handler(handler);
+    // }
 
     #[must_use]
     pub fn subscribed_quotes(&self) -> Vec<InstrumentId> {
