+++ crates/model/src/lib.rs
@@ -49,6 +49,7 @@ pub mod macros;
 pub mod orderbook;
 pub mod orders;
 pub mod position;
+pub mod reports;
 pub mod types;
 pub mod venues;
